+++ MaMar/KKMC.h
@@ -23,25 +23,25 @@ using namespace std;
 extern "C" void kk2f_initialize_(double xpar[]);
 extern "C" void kk2f_make_();
 extern "C" void kk2f_finalize_();
-extern "C" void kk2f_print1_(    const long&);
-extern "C" void kk2f_fort_open_( const long&, char*, long);
-extern "C" void kk2f_fort_close_(const long&);
+extern "C" void kk2f_print1_(    const int&);
+extern "C" void kk2f_fort_open_( const int&, char*, int);
+extern "C" void kk2f_fort_close_(const int&);
 ///////////////////////////////////////////////////////////////////////////////
 extern "C" void kk2f_getwt_(     double&, double&);
 extern "C" void kk2f_getxsecmc_( double&, double&);
 extern "C" void kk2f_getbeams_(   double [], double []);
 extern "C" void kk2f_getfermions_(double [], double []);
-extern "C" void kk2f_getnphot_(  long&);
-extern "C" void kk2f_getphoton1_( long&, double []);
+extern "C" void kk2f_getnphot_(  int&);
+extern "C" void kk2f_getphoton1_( int&, double []);
 extern "C" void kk2f_getprimanorma_( double&, const int&);
 extern "C" void kk2f_getxsnormpb_( double&, double&);
-extern "C" void kk2f_getwtalter_( long&, double&);
+extern "C" void kk2f_getwtalter_( int&, double&);
 ///////////////////////////////////////////////////////////////////////////////
 extern "C" void karlud_getfermions_(double [], double []);
 ///////////////////////////////////////////////////////////////////////////////
-extern "C" void pylist_(const long&);
-extern "C" void pygive_(char *directive, long s1);
-extern "C" void hepevt_getkffin_(  long&);
+extern "C" void pylist_(const int&);
+extern "C" void pygive_(char *directive, int s1);
+extern "C" void hepevt_getkffin_(  int&);
 //
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -49,10 +49,10 @@ extern "C" void hepevt_getkffin_(  long&);
 class KKMC{
 //class KoralwMaker {
  public:
-  long      m_NevTot;         //   total numer of events to be generated
-  long      m_EvenCounter;    //   event serial counter
+  int      m_NevTot;         //   total numer of events to be generated
+  int      m_EvenCounter;    //   event serial counter
   double    m_ypar[10001];    //   xpar input/output aprameters of koralw
-  long      m_out;            //   output unit number
+  int      m_out;            //   output unit number
  public:
 //------ header of event_lu constructor -------
   KKMC(){;};
@@ -66,27 +66,27 @@ class KKMC{
   void Finalize(double&,  double&);
 //------ read input data
   void Make();
-//  void ReadData( long &ntot );
+//  void ReadData( int &ntot );
   void Print1();
 ///////////////////////////////////////////////////////////////////////////////
-  long GetPyNpart();
-  void GetPyParticle( const long, PartLund &Event);
+  int GetPyNpart();
+  void GetPyParticle( const int, PartLund &Event);
 ///////////////////////////////////////////////////////////////////////////////
   void GetWt( double&,  double&);
   void GetBeams(    TLorentzVector&,  TLorentzVector&);
   void GetFermions( TLorentzVector&,  TLorentzVector&);
-  void GetPhoton1(const long, TLorentzVector&);
+  void GetPhoton1(const int, TLorentzVector&);
   void GetFermKarlud( TLorentzVector &,  TLorentzVector &);
-//  void GetPhoton1(long & , TLorentzVector&);
-  void GetNphot(  long &);
+//  void GetPhoton1(int & , TLorentzVector&);
+  void GetNphot(  int &);
   void GetXsecMC( double &,  double &);
-  void GetPrimaNorma( double &,  long &);
-  double GetWtAlter(const long );
-//  double GetWtAlter(long &);
+  void GetPrimaNorma( double &,  int &);
+  double GetWtAlter(const int );
+//  double GetWtAlter(int &);
 ///////////////////////////////////////////////////////////////////////////////
-  void PyList(long);
+  void PyList(int);
   void PyGive(char *directive);
-  void GetKFfin(  long &);
+  void GetKFfin(  int &);
 };
 ////////////////////////////////////////////////////////////////////////////
 #endif
