+++ MaMar/ROBOL2.cxx
@@ -15,11 +15,11 @@
 //      *************** temporary entries from KKMC ****************
 //      SUBROUTINE KarLud_GetVVxx(vv,x1,x2)
 extern "C" void  karlud_getvvxx_(double&, double&, double&);
-extern "C" void  pyhepc_(long&);
-extern "C" void  photos_(long&);
+extern "C" void  pyhepc_(int&);
+extern "C" void  photos_(int&);
 extern "C" void  phoini_();
-extern "C" void  hepevt_setphotosflagtrue_(long&);
-extern "C" void  hepevt_getnhep_(long&);
+extern "C" void  hepevt_setphotosflagtrue_(int&);
+extern "C" void  hepevt_getnhep_(int&);
 ///////////////////////////////////////////////////////////////////////////////
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -36,7 +36,7 @@ void ROBOL2::Initialize(long &NevTot)
   double ypar[jmax];
   for(int j=0;j<jmax;j++) ypar[j]=m_xpar[j+1];    // ypar has c++ numbering
   //
-  NevTot = (long)m_xpar[0];                       // NevTot hidden in xpar[0] !!!
+  NevTot = (int)m_xpar[0];                       // NevTot hidden in xpar[0] !!!
   KKMC_generator = new KKMC();
   KKMC_generator->Initialize(ypar);
   cout<<"ROBOL2::Initialize:  NevTot = "<<NevTot<<endl;
@@ -122,7 +122,7 @@ void ROBOL2::KKMC_NORMA()
   // Transfer normalization Record of KKMC to local histogram.
   // For later use in re-normalizing histostograms
   //
-  double XsPrim; long NevPrim;
+  double XsPrim; int NevPrim;
   KKMC_generator->GetPrimaNorma(XsPrim, NevPrim);
   HST_KKMC_NORMA->SetBinContent(0,XsPrim*NevPrim);
   HST_KKMC_NORMA->SetEntries(NevPrim);
@@ -150,7 +150,7 @@ void ROBOL2::Production(long &iEvent)
   KKMC_generator->GetFermions(m_pfer1,m_pfer2);
   KKMC_generator->GetNphot(m_Nphot);                  // photon multiplicity
   TLorentzVector VSumPhot;    // By default all components are initialized by zero. 
-  long iphot,iphot1;
+  int iphot,iphot1;
   for(iphot=0;iphot<m_Nphot;iphot++){
     KKMC_generator->GetPhoton1(iphot+1,m_phot[iphot]);  // photon 4-momenta
     VSumPhot+= m_phot[iphot];
@@ -186,7 +186,7 @@ void ROBOL2::Production(long &iEvent)
   int    TrigPho=1;       /// =1 for accepted, =0 for rejected
   ///**************************************************
   /// **** fermion inentification ****
-  long KF=0, NuYes=0;
+  int KF=0, NuYes=0;
 //  KKMC_generator->GetKFfin(KF); /// does not work
   if( PartFindStable( 12) ) KF=12;  /// electron neutrino found
   if( PartFindStable( 14) ) KF=14;  /// mu neutrino found
@@ -332,7 +332,7 @@ void ROBOL2::Finalize()
   hstN_Mff->Add(hstN_Mff, Fact);
   // *********************************************************************
   // **** alternatively HST_KKMC_NORMA is used at later stage (plotting)
-  long   NevPrim = HST_KKMC_NORMA->GetEntries();
+  int   NevPrim = HST_KKMC_NORMA->GetEntries();
   double XsPrima = HST_KKMC_NORMA->GetBinContent(0)/NevPrim;
   cout << "HST_KKMC_NORMA: XsPrima [nb] = "<< XsPrima << " NevPrim= "<< NevPrim <<endl;
   // *********************************************************************
@@ -360,35 +360,35 @@ void ROBOL2::PartImport(){
     cout<<"++++ ROBOL2::Production: STOP m_Npart= "<<m_Npart<<endl;
     exit(5);
   }
-  for(long j=0; j<m_Npart;j++){
+  for(int j=0; j<m_Npart;j++){
     KKMC_generator->GetPyParticle( j, m_Event[j]);  // import one particle
     //m_Event[j].Print(1);
   }
 }
 ///////////////////////////////////////////////////////////////////////////////
-long ROBOL2::PartCount(const long flav){
-  long jCount=0;
-  for(long j=0; j<m_Npart;j++)
+int ROBOL2::PartCount(const int flav){
+  int jCount=0;
+  for(int j=0; j<m_Npart;j++)
     if(m_Event[j].fFlafor == flav){
       jCount++;
     }
   return jCount;
 }
 ///////////////////////////////////////////////////////////////////////////////
-long ROBOL2::PartFindAny(const long flav){
+int ROBOL2::PartFindAny(const int flav){
 // fortran numbering!!!
-  long jPosition=0;
-  for(long j=0; j<m_Npart;j++)
+  int jPosition=0;
+  for(int j=0; j<m_Npart;j++)
     if(m_Event[j].fFlafor == flav){
       jPosition=j+1; break;
     }
   return jPosition;
 }
 ///////////////////////////////////////////////////////////////////////////////
-long ROBOL2::PartFindStable(const long flav){
+int ROBOL2::PartFindStable(const int flav){
 // fortran numbering!!!
-  long jPosition=0;
-  for(long j=0; j<m_Npart;j++)
+  int jPosition=0;
+  for(int j=0; j<m_Npart;j++)
     if((m_Event[j].fStatus  == 1)&&(m_Event[j].fFlafor == flav)){
       jPosition=j+1; break;
     }
@@ -404,7 +404,7 @@ void ROBOL2::PyPrint(const int mode){
   cout<<"lser status flavor parent child1 child2";
   cout<<"                Px                Py                Pz               Ene";
   cout<<"              Mass"<<endl;
-  for(long j=0; j<m_Npart;j++){
+  for(int j=0; j<m_Npart;j++){
     m_Event[j].Print(mode);
     if(m_Event[j].fStatus == 1) Sum += m_Event[j].fMom;
   }
