+++ MaMar/ROBOL2.h
@@ -16,16 +16,16 @@
 //class ROBOL : public TNamed {
 class ROBOL2{
  public:
-  long   m_NevGen;           // event serial number
-  long   m_count1;           // auxiliary event counter (debug)
+  long  m_NevGen;           // event serial number
+  long  m_count1;           // auxiliary event counter (debug)
   double m_xpar[10001];      // complete input of KKMC run
   KKMC   *KKMC_generator;    // MC event generator
   // =============== local mirror of KKMC event =======================
   TLorentzVector m_pbea1,m_pbea2;    // initial beams
   TLorentzVector m_pfer1,m_pfer2;    // final fermions
-  long           m_Nphot;            // photon multiplicity
+  int           m_Nphot;            // photon multiplicity
   TLorentzVector m_phot[100];        // photon 4-momenta
-  long           m_Npart;            // no of particles in Lund/Pythia common
+  int           m_Npart;            // no of particles in Lund/Pythia common
   PartLund m_Event[4000];            // content of /PYJETS/
   // ============== Histograms follow =================================
   TH1D   *HST_KKMC_NORMA;    // special histo with KKMC normalization & xpar
@@ -80,9 +80,9 @@ class ROBOL2{
   void KKMC_NORMA();
   void Finalize();
   void PartImport();
-  long PartCount(  const long);
-  long PartFindAny(const long);
-  long PartFindStable(const long);
+  int PartCount(  const int);
+  int PartFindAny(const int);
+  int PartFindStable(const int);
   void PyPrint(const int );
   void MomPrint( TLorentzVector&);
   void ReaData(char DiskFile[], int imax, double xpar[]);
